refactor: escaping query
